Editor Modifications
Modified the editor so it produces a new grid every time instead of trying to replace the old one. The user must now provide the grid they want the editor to modify. Also modified the editor so that the Add Tile modifier won't allow tiles to be placed at the same spot on the same grid.